@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.0</version>
+        <version>3.0.8</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.cydeo</groupId>
@@ -14,7 +14,7 @@
     <name>Spring-19-REST-Security-Keycloak-LicenseApp</name>
     <description>Spring-19-REST-Security-Keycloak-LicenseApp</description>
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
     </properties>
     <dependencies>
         <dependency>
@@ -28,7 +28,7 @@
         <dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-starter-openfeign</artifactId>
-            <version>3.1.3</version>
+            <version>4.0.3</version>
         </dependency>
         <dependency>
             <groupId>org.postgresql</groupId>
